+++ internal/api/anilist/client_gen.go
@@ -17,6 +17,7 @@ type GithubGraphQLClient interface {
 	CompleteAnimeByID(ctx context.Context, id *int, interceptors ...clientv2.RequestInterceptor) (*CompleteAnimeByID, error)
 	AnimeDetailsByID(ctx context.Context, id *int, interceptors ...clientv2.RequestInterceptor) (*AnimeDetailsByID, error)
 	ListAnime(ctx context.Context, page *int, search *string, perPage *int, sort []*MediaSort, status []*MediaStatus, genres []*string, averageScoreGreater *int, season *MediaSeason, seasonYear *int, format *MediaFormat, isAdult *bool, interceptors ...clientv2.RequestInterceptor) (*ListAnime, error)
+	ListAnimeAll(ctx context.Context, page *int, search *string, perPage *int, sort []*MediaSort, status []*MediaStatus, genres []*string, averageScoreGreater *int, season *MediaSeason, seasonYear *int, format *MediaFormat, interceptors ...clientv2.RequestInterceptor) (*ListAnimeAll, error)
 	ListRecentAnime(ctx context.Context, page *int, perPage *int, airingAtGreater *int, airingAtLesser *int, notYetAired *bool, interceptors ...clientv2.RequestInterceptor) (*ListRecentAnime, error)
 	AnimeAiringSchedule(ctx context.Context, ids []*int, season *MediaSeason, seasonYear *int, previousSeason *MediaSeason, previousSeasonYear *int, nextSeason *MediaSeason, nextSeasonYear *int, interceptors ...clientv2.RequestInterceptor) (*AnimeAiringSchedule, error)
 	AnimeAiringScheduleRaw(ctx context.Context, ids []*int, interceptors ...clientv2.RequestInterceptor) (*AnimeAiringScheduleRaw, error)
@@ -29,6 +30,7 @@ type GithubGraphQLClient interface {
 	BaseMangaByID(ctx context.Context, id *int, interceptors ...clientv2.RequestInterceptor) (*BaseMangaByID, error)
 	MangaDetailsByID(ctx context.Context, id *int, interceptors ...clientv2.RequestInterceptor) (*MangaDetailsByID, error)
 	ListManga(ctx context.Context, page *int, search *string, perPage *int, sort []*MediaSort, status []*MediaStatus, genres []*string, averageScoreGreater *int, startDateGreater *string, startDateLesser *string, format *MediaFormat, countryOfOrigin *string, isAdult *bool, interceptors ...clientv2.RequestInterceptor) (*ListManga, error)
+	ListMangaAll(ctx context.Context, page *int, search *string, perPage *int, sort []*MediaSort, status []*MediaStatus, genres []*string, averageScoreGreater *int, startDateGreater *string, startDateLesser *string, format *MediaFormat, countryOfOrigin *string, interceptors ...clientv2.RequestInterceptor) (*ListMangaAll, error)
 	ViewerStats(ctx context.Context, interceptors ...clientv2.RequestInterceptor) (*ViewerStats, error)
 	StudioDetails(ctx context.Context, id *int, interceptors ...clientv2.RequestInterceptor) (*StudioDetails, error)
 	GetViewer(ctx context.Context, interceptors ...clientv2.RequestInterceptor) (*GetViewer, error)
@@ -4589,6 +4591,227 @@ func (t *ListAnime_Page) GetPageInfo() *ListAnime_Page_PageInfo {
 	return t.PageInfo
 }
 
+type ListAnimeAll_Page_PageInfo struct {
+	CurrentPage *int  "json:\"currentPage,omitempty\" graphql:\"currentPage\""
+	HasNextPage *bool "json:\"hasNextPage,omitempty\" graphql:\"hasNextPage\""
+	LastPage    *int  "json:\"lastPage,omitempty\" graphql:\"lastPage\""
+	PerPage     *int  "json:\"perPage,omitempty\" graphql:\"perPage\""
+	Total       *int  "json:\"total,omitempty\" graphql:\"total\""
+}
+
+func (t *ListAnimeAll_Page_PageInfo) GetCurrentPage() *int {
+	if t == nil {
+		t = &ListAnimeAll_Page_PageInfo{}
+	}
+	return t.CurrentPage
+}
+func (t *ListAnimeAll_Page_PageInfo) GetHasNextPage() *bool {
+	if t == nil {
+		t = &ListAnimeAll_Page_PageInfo{}
+	}
+	return t.HasNextPage
+}
+func (t *ListAnimeAll_Page_PageInfo) GetLastPage() *int {
+	if t == nil {
+		t = &ListAnimeAll_Page_PageInfo{}
+	}
+	return t.LastPage
+}
+func (t *ListAnimeAll_Page_PageInfo) GetPerPage() *int {
+	if t == nil {
+		t = &ListAnimeAll_Page_PageInfo{}
+	}
+	return t.PerPage
+}
+func (t *ListAnimeAll_Page_PageInfo) GetTotal() *int {
+	if t == nil {
+		t = &ListAnimeAll_Page_PageInfo{}
+	}
+	return t.Total
+}
+
+type ListAnimeAll_Page_Media_BaseAnime_Trailer struct {
+	ID        *string "json:\"id,omitempty\" graphql:\"id\""
+	Site      *string "json:\"site,omitempty\" graphql:\"site\""
+	Thumbnail *string "json:\"thumbnail,omitempty\" graphql:\"thumbnail\""
+}
+
+func (t *ListAnimeAll_Page_Media_BaseAnime_Trailer) GetID() *string {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_Trailer{}
+	}
+	return t.ID
+}
+func (t *ListAnimeAll_Page_Media_BaseAnime_Trailer) GetSite() *string {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_Trailer{}
+	}
+	return t.Site
+}
+func (t *ListAnimeAll_Page_Media_BaseAnime_Trailer) GetThumbnail() *string {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_Trailer{}
+	}
+	return t.Thumbnail
+}
+
+type ListAnimeAll_Page_Media_BaseAnime_Title struct {
+	English       *string "json:\"english,omitempty\" graphql:\"english\""
+	Native        *string "json:\"native,omitempty\" graphql:\"native\""
+	Romaji        *string "json:\"romaji,omitempty\" graphql:\"romaji\""
+	UserPreferred *string "json:\"userPreferred,omitempty\" graphql:\"userPreferred\""
+}
+
+func (t *ListAnimeAll_Page_Media_BaseAnime_Title) GetEnglish() *string {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_Title{}
+	}
+	return t.English
+}
+func (t *ListAnimeAll_Page_Media_BaseAnime_Title) GetNative() *string {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_Title{}
+	}
+	return t.Native
+}
+func (t *ListAnimeAll_Page_Media_BaseAnime_Title) GetRomaji() *string {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_Title{}
+	}
+	return t.Romaji
+}
+func (t *ListAnimeAll_Page_Media_BaseAnime_Title) GetUserPreferred() *string {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_Title{}
+	}
+	return t.UserPreferred
+}
+
+type ListAnimeAll_Page_Media_BaseAnime_CoverImage struct {
+	Color      *string "json:\"color,omitempty\" graphql:\"color\""
+	ExtraLarge *string "json:\"extraLarge,omitempty\" graphql:\"extraLarge\""
+	Large      *string "json:\"large,omitempty\" graphql:\"large\""
+	Medium     *string "json:\"medium,omitempty\" graphql:\"medium\""
+}
+
+func (t *ListAnimeAll_Page_Media_BaseAnime_CoverImage) GetColor() *string {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_CoverImage{}
+	}
+	return t.Color
+}
+func (t *ListAnimeAll_Page_Media_BaseAnime_CoverImage) GetExtraLarge() *string {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_CoverImage{}
+	}
+	return t.ExtraLarge
+}
+func (t *ListAnimeAll_Page_Media_BaseAnime_CoverImage) GetLarge() *string {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_CoverImage{}
+	}
+	return t.Large
+}
+func (t *ListAnimeAll_Page_Media_BaseAnime_CoverImage) GetMedium() *string {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_CoverImage{}
+	}
+	return t.Medium
+}
+
+type ListAnimeAll_Page_Media_BaseAnime_StartDate struct {
+	Day   *int "json:\"day,omitempty\" graphql:\"day\""
+	Month *int "json:\"month,omitempty\" graphql:\"month\""
+	Year  *int "json:\"year,omitempty\" graphql:\"year\""
+}
+
+func (t *ListAnimeAll_Page_Media_BaseAnime_StartDate) GetDay() *int {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_StartDate{}
+	}
+	return t.Day
+}
+func (t *ListAnimeAll_Page_Media_BaseAnime_StartDate) GetMonth() *int {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_StartDate{}
+	}
+	return t.Month
+}
+func (t *ListAnimeAll_Page_Media_BaseAnime_StartDate) GetYear() *int {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_StartDate{}
+	}
+	return t.Year
+}
+
+type ListAnimeAll_Page_Media_BaseAnime_EndDate struct {
+	Day   *int "json:\"day,omitempty\" graphql:\"day\""
+	Month *int "json:\"month,omitempty\" graphql:\"month\""
+	Year  *int "json:\"year,omitempty\" graphql:\"year\""
+}
+
+func (t *ListAnimeAll_Page_Media_BaseAnime_EndDate) GetDay() *int {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_EndDate{}
+	}
+	return t.Day
+}
+func (t *ListAnimeAll_Page_Media_BaseAnime_EndDate) GetMonth() *int {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_EndDate{}
+	}
+	return t.Month
+}
+func (t *ListAnimeAll_Page_Media_BaseAnime_EndDate) GetYear() *int {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_EndDate{}
+	}
+	return t.Year
+}
+
+type ListAnimeAll_Page_Media_BaseAnime_NextAiringEpisode struct {
+	AiringAt        int "json:\"airingAt\" graphql:\"airingAt\""
+	Episode         int "json:\"episode\" graphql:\"episode\""
+	TimeUntilAiring int "json:\"timeUntilAiring\" graphql:\"timeUntilAiring\""
+}
+
+func (t *ListAnimeAll_Page_Media_BaseAnime_NextAiringEpisode) GetAiringAt() int {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_NextAiringEpisode{}
+	}
+	return t.AiringAt
+}
+func (t *ListAnimeAll_Page_Media_BaseAnime_NextAiringEpisode) GetEpisode() int {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_NextAiringEpisode{}
+	}
+	return t.Episode
+}
+func (t *ListAnimeAll_Page_Media_BaseAnime_NextAiringEpisode) GetTimeUntilAiring() int {
+	if t == nil {
+		t = &ListAnimeAll_Page_Media_BaseAnime_NextAiringEpisode{}
+	}
+	return t.TimeUntilAiring
+}
+
+type ListAnimeAll_Page struct {
+	Media    []*BaseAnime                "json:\"media,omitempty\" graphql:\"media\""
+	PageInfo *ListAnimeAll_Page_PageInfo "json:\"pageInfo,omitempty\" graphql:\"pageInfo\""
+}
+
+func (t *ListAnimeAll_Page) GetMedia() []*BaseAnime {
+	if t == nil {
+		t = &ListAnimeAll_Page{}
+	}
+	return t.Media
+}
+func (t *ListAnimeAll_Page) GetPageInfo() *ListAnimeAll_Page_PageInfo {
+	if t == nil {
+		t = &ListAnimeAll_Page{}
+	}
+	return t.PageInfo
+}
+
 type ListRecentAnime_Page_PageInfo struct {
 	CurrentPage *int  "json:\"currentPage,omitempty\" graphql:\"currentPage\""
 	HasNextPage *bool "json:\"hasNextPage,omitempty\" graphql:\"hasNextPage\""
@@ -6744,6 +6967,177 @@ func (t *ListManga_Page) GetPageInfo() *ListManga_Page_PageInfo {
 	return t.PageInfo
 }
 
+type ListMangaAll_Page_PageInfo struct {
+	CurrentPage *int  "json:\"currentPage,omitempty\" graphql:\"currentPage\""
+	HasNextPage *bool "json:\"hasNextPage,omitempty\" graphql:\"hasNextPage\""
+	LastPage    *int  "json:\"lastPage,omitempty\" graphql:\"lastPage\""
+	PerPage     *int  "json:\"perPage,omitempty\" graphql:\"perPage\""
+	Total       *int  "json:\"total,omitempty\" graphql:\"total\""
+}
+
+func (t *ListMangaAll_Page_PageInfo) GetCurrentPage() *int {
+	if t == nil {
+		t = &ListMangaAll_Page_PageInfo{}
+	}
+	return t.CurrentPage
+}
+func (t *ListMangaAll_Page_PageInfo) GetHasNextPage() *bool {
+	if t == nil {
+		t = &ListMangaAll_Page_PageInfo{}
+	}
+	return t.HasNextPage
+}
+func (t *ListMangaAll_Page_PageInfo) GetLastPage() *int {
+	if t == nil {
+		t = &ListMangaAll_Page_PageInfo{}
+	}
+	return t.LastPage
+}
+func (t *ListMangaAll_Page_PageInfo) GetPerPage() *int {
+	if t == nil {
+		t = &ListMangaAll_Page_PageInfo{}
+	}
+	return t.PerPage
+}
+func (t *ListMangaAll_Page_PageInfo) GetTotal() *int {
+	if t == nil {
+		t = &ListMangaAll_Page_PageInfo{}
+	}
+	return t.Total
+}
+
+type ListMangaAll_Page_Media_BaseManga_Title struct {
+	English       *string "json:\"english,omitempty\" graphql:\"english\""
+	Native        *string "json:\"native,omitempty\" graphql:\"native\""
+	Romaji        *string "json:\"romaji,omitempty\" graphql:\"romaji\""
+	UserPreferred *string "json:\"userPreferred,omitempty\" graphql:\"userPreferred\""
+}
+
+func (t *ListMangaAll_Page_Media_BaseManga_Title) GetEnglish() *string {
+	if t == nil {
+		t = &ListMangaAll_Page_Media_BaseManga_Title{}
+	}
+	return t.English
+}
+func (t *ListMangaAll_Page_Media_BaseManga_Title) GetNative() *string {
+	if t == nil {
+		t = &ListMangaAll_Page_Media_BaseManga_Title{}
+	}
+	return t.Native
+}
+func (t *ListMangaAll_Page_Media_BaseManga_Title) GetRomaji() *string {
+	if t == nil {
+		t = &ListMangaAll_Page_Media_BaseManga_Title{}
+	}
+	return t.Romaji
+}
+func (t *ListMangaAll_Page_Media_BaseManga_Title) GetUserPreferred() *string {
+	if t == nil {
+		t = &ListMangaAll_Page_Media_BaseManga_Title{}
+	}
+	return t.UserPreferred
+}
+
+type ListMangaAll_Page_Media_BaseManga_CoverImage struct {
+	Color      *string "json:\"color,omitempty\" graphql:\"color\""
+	ExtraLarge *string "json:\"extraLarge,omitempty\" graphql:\"extraLarge\""
+	Large      *string "json:\"large,omitempty\" graphql:\"large\""
+	Medium     *string "json:\"medium,omitempty\" graphql:\"medium\""
+}
+
+func (t *ListMangaAll_Page_Media_BaseManga_CoverImage) GetColor() *string {
+	if t == nil {
+		t = &ListMangaAll_Page_Media_BaseManga_CoverImage{}
+	}
+	return t.Color
+}
+func (t *ListMangaAll_Page_Media_BaseManga_CoverImage) GetExtraLarge() *string {
+	if t == nil {
+		t = &ListMangaAll_Page_Media_BaseManga_CoverImage{}
+	}
+	return t.ExtraLarge
+}
+func (t *ListMangaAll_Page_Media_BaseManga_CoverImage) GetLarge() *string {
+	if t == nil {
+		t = &ListMangaAll_Page_Media_BaseManga_CoverImage{}
+	}
+	return t.Large
+}
+func (t *ListMangaAll_Page_Media_BaseManga_CoverImage) GetMedium() *string {
+	if t == nil {
+		t = &ListMangaAll_Page_Media_BaseManga_CoverImage{}
+	}
+	return t.Medium
+}
+
+type ListMangaAll_Page_Media_BaseManga_StartDate struct {
+	Day   *int "json:\"day,omitempty\" graphql:\"day\""
+	Month *int "json:\"month,omitempty\" graphql:\"month\""
+	Year  *int "json:\"year,omitempty\" graphql:\"year\""
+}
+
+func (t *ListMangaAll_Page_Media_BaseManga_StartDate) GetDay() *int {
+	if t == nil {
+		t = &ListMangaAll_Page_Media_BaseManga_StartDate{}
+	}
+	return t.Day
+}
+func (t *ListMangaAll_Page_Media_BaseManga_StartDate) GetMonth() *int {
+	if t == nil {
+		t = &ListMangaAll_Page_Media_BaseManga_StartDate{}
+	}
+	return t.Month
+}
+func (t *ListMangaAll_Page_Media_BaseManga_StartDate) GetYear() *int {
+	if t == nil {
+		t = &ListMangaAll_Page_Media_BaseManga_StartDate{}
+	}
+	return t.Year
+}
+
+type ListMangaAll_Page_Media_BaseManga_EndDate struct {
+	Day   *int "json:\"day,omitempty\" graphql:\"day\""
+	Month *int "json:\"month,omitempty\" graphql:\"month\""
+	Year  *int "json:\"year,omitempty\" graphql:\"year\""
+}
+
+func (t *ListMangaAll_Page_Media_BaseManga_EndDate) GetDay() *int {
+	if t == nil {
+		t = &ListMangaAll_Page_Media_BaseManga_EndDate{}
+	}
+	return t.Day
+}
+func (t *ListMangaAll_Page_Media_BaseManga_EndDate) GetMonth() *int {
+	if t == nil {
+		t = &ListMangaAll_Page_Media_BaseManga_EndDate{}
+	}
+	return t.Month
+}
+func (t *ListMangaAll_Page_Media_BaseManga_EndDate) GetYear() *int {
+	if t == nil {
+		t = &ListMangaAll_Page_Media_BaseManga_EndDate{}
+	}
+	return t.Year
+}
+
+type ListMangaAll_Page struct {
+	Media    []*BaseManga                "json:\"media,omitempty\" graphql:\"media\""
+	PageInfo *ListMangaAll_Page_PageInfo "json:\"pageInfo,omitempty\" graphql:\"pageInfo\""
+}
+
+func (t *ListMangaAll_Page) GetMedia() []*BaseManga {
+	if t == nil {
+		t = &ListMangaAll_Page{}
+	}
+	return t.Media
+}
+func (t *ListMangaAll_Page) GetPageInfo() *ListMangaAll_Page_PageInfo {
+	if t == nil {
+		t = &ListMangaAll_Page{}
+	}
+	return t.PageInfo
+}
+
 type ViewerStats_Viewer_Statistics_Anime_Studios_UserStudioStats_Studio struct {
 	ID                int    "json:\"id\" graphql:\"id\""
 	IsAnimationStudio bool   "json:\"isAnimationStudio\" graphql:\"isAnimationStudio\""
@@ -7355,6 +7749,17 @@ func (t *ListAnime) GetPage() *ListAnime_Page {
 	return t.Page
 }
 
+type ListAnimeAll struct {
+	Page *ListAnimeAll_Page "json:\"Page,omitempty\" graphql:\"Page\""
+}
+
+func (t *ListAnimeAll) GetPage() *ListAnimeAll_Page {
+	if t == nil {
+		t = &ListAnimeAll{}
+	}
+	return t.Page
+}
+
 type ListRecentAnime struct {
 	Page *ListRecentAnime_Page "json:\"Page,omitempty\" graphql:\"Page\""
 }
@@ -7511,6 +7916,17 @@ type ListManga struct {
 func (t *ListManga) GetPage() *ListManga_Page {
 	if t == nil {
 		t = &ListManga{}
+	}
+	return t.Page
+}
+
+type ListMangaAll struct {
+	Page *ListMangaAll_Page "json:\"Page,omitempty\" graphql:\"Page\""
+}
+
+func (t *ListMangaAll) GetPage() *ListMangaAll_Page {
+	if t == nil {
+		t = &ListMangaAll{}
 	}
 	return t.Page
 }
@@ -8480,6 +8896,99 @@ func (c *Client) ListAnime(ctx context.Context, page *int, search *string, perPa
 	return &res, nil
 }
 
+const ListAnimeAllDocument = `query ListAnimeAll ($page: Int, $search: String, $perPage: Int, $sort: [MediaSort], $status: [MediaStatus], $genres: [String], $averageScore_greater: Int, $season: MediaSeason, $seasonYear: Int, $format: MediaFormat) {
+	Page(page: $page, perPage: $perPage) {
+		pageInfo {
+			hasNextPage
+			total
+			perPage
+			currentPage
+			lastPage
+		}
+		media(type: ANIME, search: $search, sort: $sort, status_in: $status, format: $format, genre_in: $genres, averageScore_greater: $averageScore_greater, season: $season, seasonYear: $seasonYear, format_not: MUSIC) {
+			... baseAnime
+		}
+	}
+}
+fragment baseAnime on Media {
+	id
+	idMal
+	siteUrl
+	status(version: 2)
+	season
+	type
+	format
+	seasonYear
+	bannerImage
+	episodes
+	synonyms
+	isAdult
+	countryOfOrigin
+	meanScore
+	description
+	genres
+	duration
+	trailer {
+		id
+		site
+		thumbnail
+	}
+	title {
+		userPreferred
+		romaji
+		english
+		native
+	}
+	coverImage {
+		extraLarge
+		large
+		medium
+		color
+	}
+	startDate {
+		year
+		month
+		day
+	}
+	endDate {
+		year
+		month
+		day
+	}
+	nextAiringEpisode {
+		airingAt
+		timeUntilAiring
+		episode
+	}
+}
+`
+
+func (c *Client) ListAnimeAll(ctx context.Context, page *int, search *string, perPage *int, sort []*MediaSort, status []*MediaStatus, genres []*string, averageScoreGreater *int, season *MediaSeason, seasonYear *int, format *MediaFormat, interceptors ...clientv2.RequestInterceptor) (*ListAnimeAll, error) {
+	vars := map[string]any{
+		"page":                 page,
+		"search":               search,
+		"perPage":              perPage,
+		"sort":                 sort,
+		"status":               status,
+		"genres":               genres,
+		"averageScore_greater": averageScoreGreater,
+		"season":               season,
+		"seasonYear":           seasonYear,
+		"format":               format,
+	}
+
+	var res ListAnimeAll
+	if err := c.Client.Post(ctx, "ListAnimeAll", ListAnimeAllDocument, &res, vars, interceptors...); err != nil {
+		if c.Client.ParseDataWhenErrors {
+			return &res, err
+		}
+
+		return nil, err
+	}
+
+	return &res, nil
+}
+
 const ListRecentAnimeDocument = `query ListRecentAnime ($page: Int, $perPage: Int, $airingAt_greater: Int, $airingAt_lesser: Int, $notYetAired: Boolean = false) {
 	Page(page: $page, perPage: $perPage) {
 		pageInfo {
@@ -9263,6 +9772,89 @@ func (c *Client) ListManga(ctx context.Context, page *int, search *string, perPa
 	return &res, nil
 }
 
+const ListMangaAllDocument = `query ListMangaAll ($page: Int, $search: String, $perPage: Int, $sort: [MediaSort], $status: [MediaStatus], $genres: [String], $averageScore_greater: Int, $startDate_greater: FuzzyDateInt, $startDate_lesser: FuzzyDateInt, $format: MediaFormat, $countryOfOrigin: CountryCode) {
+	Page(page: $page, perPage: $perPage) {
+		pageInfo {
+			hasNextPage
+			total
+			perPage
+			currentPage
+			lastPage
+		}
+		media(type: MANGA, countryOfOrigin: $countryOfOrigin, search: $search, sort: $sort, status_in: $status, format: $format, genre_in: $genres, averageScore_greater: $averageScore_greater, startDate_greater: $startDate_greater, startDate_lesser: $startDate_lesser, format_not: NOVEL) {
+			... baseManga
+		}
+	}
+}
+fragment baseManga on Media {
+	id
+	idMal
+	siteUrl
+	status(version: 2)
+	season
+	type
+	format
+	bannerImage
+	chapters
+	volumes
+	synonyms
+	isAdult
+	countryOfOrigin
+	meanScore
+	description
+	genres
+	title {
+		userPreferred
+		romaji
+		english
+		native
+	}
+	coverImage {
+		extraLarge
+		large
+		medium
+		color
+	}
+	startDate {
+		year
+		month
+		day
+	}
+	endDate {
+		year
+		month
+		day
+	}
+}
+`
+
+func (c *Client) ListMangaAll(ctx context.Context, page *int, search *string, perPage *int, sort []*MediaSort, status []*MediaStatus, genres []*string, averageScoreGreater *int, startDateGreater *string, startDateLesser *string, format *MediaFormat, countryOfOrigin *string, interceptors ...clientv2.RequestInterceptor) (*ListMangaAll, error) {
+	vars := map[string]any{
+		"page":                 page,
+		"search":               search,
+		"perPage":              perPage,
+		"sort":                 sort,
+		"status":               status,
+		"genres":               genres,
+		"averageScore_greater": averageScoreGreater,
+		"startDate_greater":    startDateGreater,
+		"startDate_lesser":     startDateLesser,
+		"format":               format,
+		"countryOfOrigin":      countryOfOrigin,
+	}
+
+	var res ListMangaAll
+	if err := c.Client.Post(ctx, "ListMangaAll", ListMangaAllDocument, &res, vars, interceptors...); err != nil {
+		if c.Client.ParseDataWhenErrors {
+			return &res, err
+		}
+
+		return nil, err
+	}
+
+	return &res, nil
+}
+
 const ViewerStatsDocument = `query ViewerStats {
 	Viewer {
 		statistics {
@@ -9523,6 +10115,7 @@ var DocumentOperationNames = map[string]string{
 	CompleteAnimeByIDDocument:            "CompleteAnimeById",
 	AnimeDetailsByIDDocument:             "AnimeDetailsById",
 	ListAnimeDocument:                    "ListAnime",
+	ListAnimeAllDocument:                 "ListAnimeAll",
 	ListRecentAnimeDocument:              "ListRecentAnime",
 	AnimeAiringScheduleDocument:          "AnimeAiringSchedule",
 	AnimeAiringScheduleRawDocument:       "AnimeAiringScheduleRaw",
@@ -9535,6 +10128,7 @@ var DocumentOperationNames = map[string]string{
 	BaseMangaByIDDocument:                "BaseMangaById",
 	MangaDetailsByIDDocument:             "MangaDetailsById",
 	ListMangaDocument:                    "ListManga",
+	ListMangaAllDocument:                 "ListMangaAll",
 	ViewerStatsDocument:                  "ViewerStats",
 	StudioDetailsDocument:                "StudioDetails",
 	GetViewerDocument:                    "GetViewer",